requirements.txt
numpy>=1.21.0,<2.0
opencv-python==4.8.1.78
Pillow==10.0.1
tqdm==4.66.1 